Changed the directory structure of frontend/src
